+++ packages/tldraw/src/index.ts
@@ -134,7 +134,8 @@ export {
 	getArrowTerminalsInArrowSpace,
 	type TLArrowBindings,
 } from './lib/shapes/arrow/shared'
-export { BookmarkShapeUtil, createBookmarkFromUrl } from './lib/shapes/bookmark/BookmarkShapeUtil'
+export { createBookmarkFromUrl } from './lib/shapes/bookmark/bookmarks'
+export { BookmarkShapeUtil } from './lib/shapes/bookmark/BookmarkShapeUtil'
 export { DrawShapeTool } from './lib/shapes/draw/DrawShapeTool'
 export { DrawShapeUtil, type DrawShapeOptions } from './lib/shapes/draw/DrawShapeUtil'
 export { EmbedShapeUtil } from './lib/shapes/embed/EmbedShapeUtil'
