+++ packages/tldraw/src/lib/ui/context/actions.tsx
@@ -20,13 +20,12 @@ import {
 	approximately,
 	compact,
 	createShapeId,
-	deferAsyncEffects,
 	kickoutOccludedShapes,
 	openWindow,
 	useMaybeEditor,
 } from '@tldraw/editor'
 import * as React from 'react'
-import { createBookmarkFromUrl } from '../../shapes/bookmark/BookmarkShapeUtil'
+import { createBookmarkFromUrl } from '../../shapes/bookmark/bookmarks'
 import { fitFrameToContent, removeFrame } from '../../utils/frames/frames'
 import { generateShapeAnnouncementMessage } from '../components/A11y'
 import { EditLinkDialog } from '../components/EditLinkDialog'
@@ -401,30 +400,28 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 
 					const creationPromises: Promise<Result<any, any>>[] = []
 
-					await deferAsyncEffects(async () => {
-						for (const shape of shapes) {
-							if (!shape || !editor.isShapeOfType<TLEmbedShape>(shape, 'embed') || !shape.props.url)
-								continue
+					for (const shape of shapes) {
+						if (!shape || !editor.isShapeOfType<TLEmbedShape>(shape, 'embed') || !shape.props.url)
+							continue
 
-							const center = editor.getShapePageBounds(shape)?.center
+						const center = editor.getShapePageBounds(shape)?.center
 
-							if (!center) continue
-							editor.deleteShapes([shape.id])
+						if (!center) continue
+						editor.deleteShapes([shape.id])
 
-							creationPromises.push(
-								createBookmarkFromUrl(editor, { url: shape.props.url, center }).then((res) => {
-									if (!res.ok) {
-										throw new Error(res.error)
-									}
-									return res
-								})
-							)
-						}
+						creationPromises.push(
+							createBookmarkFromUrl(editor, { url: shape.props.url, center }).then((res) => {
+								if (!res.ok) {
+									throw new Error(res.error)
+								}
+								return res
+							})
+						)
+					}
 
-						await Promise.all(creationPromises).catch((error) => {
-							editor.bailToMark(markId)
-							console.error(error)
-						})
+					await Promise.all(creationPromises).catch((error) => {
+						editor.bailToMark(markId)
+						console.error(error)
 					})
 				},
 			},
